requirements.txt
# Core ML/RL libraries
torch>=2.0.0
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.0.0
tianshou>=0.5.0

# Data handling
h5py>=3.8.0
pyarrow>=12.0.0
polars>=0.18.0

# Market data APIs
requests>=2.31.0
websockets>=11.0.0
aiohttp>=3.8.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Performance
numba>=0.57.0
cython>=3.0.0

# Async programming
asyncio-mqtt>=0.13.0
